requirements.txt
future==0.16.0
oauthlib==2.0.1
olefile==0.44
Pillow==4.0.0
python-twitter==3.2
requests==2.12.4
requests-oauthlib==0.7.0
